(#8312) fix zypper package upgrade detection
The zypper package manager uses a different layout
for version information output than yum (although
both are designed to work with RPM based systems).

So far, the zypper provider was parsing the wrong
version fields and failed to notice packages that
could be updated when ensuring "latest". With this
patch, the "latest" target is working on zypper systems
like SLES and OpenSUSE.