(maint) cron: change the integration tests to rely on the compiler
The artificial catalogs that had previously been used for integration
testing are prone to issues, e.g. the containment of generated resources
would not work and lead to invalid catalogs and failing tests.

Instead of mock resources in patchwork catalogs, use actual manifests
and the compiler to end up with more realistic test cases.

WIP: Sadly, the catalogs don't lead to the expected results when applied.