(PUP-2790) Use telinit to detect upstart version.
During installations, initctl is overridden by a script that does nothing except displaying a message. This makes the original provider misbehave when forced to be used. telinit is not diverted and provides the same version information.
Still, for anything but version information, start/stop/restart etc. should remain the same, as it is actually not wanted to start a service during this stage of installation.